Wybor

Description

Console line fuzzy search as a library.

Note: It's probably better to toy with the library through the wybor executable, but it's still possible to use it from ghci, albeit it would not be pretty

First, we need a bit of setup. Both extensions aren't strictly necessary but they make our life considerably easier. Imports, on the other hand, are essential.

>>> :set -XOverloadedStrings
>>> :set -XOverloadedList
>>> import Control.Lens
>>> import Wybor

So, this starts a sligtly customized selection process (more settings are avaliable, please see HasWybor class):

>>> select (fromTexts ["foo", "bar", "baz"] & prefix .~ "λ> ")
λ>
foo
bar
baz

This new prompt (λ>) expects us to enter symbols to narrow the list of outcomes, e.g.:

λ> b
bar
baz

λ> bz
baz

At this point there is only one acceptable outcome, so we press Enter and end up with:

Right (Just "baz")

Of course, we can press Enter at any point of the process, selecting the focused alternative (marked bold here). Other hotkeys:

C-j selects the focused alternative (like Enter)

C-u clears the line

C-w deletes last word

C-h deletes last character (as does Backspace)

C-n focuses the next alternative

C-p focuses the previous alternative

C-d aborts the selection

Documentation

data Wybor a

The description of the alternative choices, see HasWybor

class HasWybor t a | t -> a where

A bunch of lenses to pick and configure Wybor

Minimal complete definition

wybor

Methods

wybor :: Lens' t (Wybor a)

alternatives :: Lens' t (NonEmpty (Text, a))

Alternative choices

visible :: Lens' t Int

How many alternative choices to show at once?

height :: Lens' t Int

How many lines every alternative takes on the screen?

prefix :: Lens' t Text

Prompt prefix

initial :: Lens' t Text

Initial search string

fromAssoc :: NonEmpty (Text, a) -> Wybor a

Construct Wybor from the nonempty list of key-value pairs

fromTexts :: NonEmpty Text -> Wybor Text

Construct Wybor from the nonempty list of strings.

The strings are used both as keys and values

select :: HasWybor t a => t -> IO (Either TTYException (Maybe a))

Select an item from Wybor once.

The user can interrupt the process with C-d and then you get Nothing. Exceptions result in Left _

selections :: (HasWybor t a, MonadResource m) => t -> Source m a

Continuously select items from Wybor. Also see select

data TTYException

Exceptions thrown while manipulating /dev/tty device

Constructors

TTYIOException IOException
Generic I/O exception

TTYNotATTY
Not-a-TTY responded with nonsense to some query

TTYConfigurationError FilePath [String] ExitCode
Failure to configure TTY with stty